debug successfully

+++ dsshaper.h
@@ -67,19 +67,19 @@ extern int last_ack_update_flag;
 
 
 /*for update_deqrate*/
-extern u64 flow_peak; //for the control peak 
+extern int flow_peak; //for the control peak 
 extern int ntrans_; // record the flow_peak change times
 extern struct timespec delay_sum_;
 extern int pktsize_sum_; //bit
 extern struct timespec checkInterval_;
 extern struct timespec checktime_; //last time update peak
 extern int alpha_; //%
-extern u64 rate_avg_; // bits/s
+extern int rate_avg_; // bits/s
 extern int delay_avg_; //us
 extern int switchOn_ ;
-extern u64 delay_optimal_;//us
-extern u64 fix_peak ; //bits/s
-extern u64 flow_peak ; // bits/s
+extern int delay_optimal_;//us
+extern int fix_peak ; //bits/s
+extern int flow_peak ; // bits/s
 ///int beta_ ; //bits/s
 //int burst_size_; //bits
 //int deltaIncrease_ ; //bits/s
@@ -207,6 +207,7 @@ struct packet_dsshaper
 
 //struct timer_list a_timer;
 #endif
+
